additionally save monitor dataset to the streak file streak values are now (long) doubles, leading to a higher precision especially in the monitor dataset add documentation for HDFCryst class

+++ include/functions.h
@@ -7,7 +7,7 @@
 #include <string>
 
 
-typedef std::vector<long> layerdata_row;
+typedef std::vector<long double> layerdata_row;
 typedef std::vector<layerdata_row> layerdata;
 
 /// remove whitespace from the beginning of a string
